Изменения в реализации
