+++ zhalilov.rail/T2/dataStruct.hpp
@@ -10,7 +10,16 @@ namespace zhalilov
     double key1;
     long long key2;
     std::string key3;
+
+    bool operator<(const DataStruct &other) const;
+    bool operator>(const DataStruct &other) const;
+    bool operator==(const DataStruct &other) const;
+    bool operator<=(const DataStruct &other) const;
+    bool operator>=(const DataStruct &other) const;
   };
+
+  std::istream &operator>>(std::istream &in, DataStruct &data);
+  std::ostream &operator<<(std::ostream &in, const DataStruct &data);
 }
 
 #endif
