--- zhalilov.rail/T2/outputData.hpp
@@ -1,12 +1,0 @@
-#ifndef OUTPUTDATA_HPP
-#define OUTPUTDATA_HPP
-
-#include <iosfwd>
-
-namespace zhalilov
-{
-  struct DataStruct;
-  std::ostream &operator<<(std::ostream &out, const DataStruct &data);
-}
-
-#endif
